wip: NLP template reorganization
