+++ gabarit/template_nlp/nlp_project/package_name/models_training/model_aggregation.py
@@ -29,9 +29,9 @@ import pandas as pd
 import dill as pickle
 from typing import Callable, Union, Tuple, Type
 
-from {{package_name}} import utils
-from {{package_name}}.models_training import utils_models
-from {{package_name}}.models_training.model_class import ModelClass
+from .. import utils
+from . import utils_models
+from .model_class import ModelClass
 
 
 def proba_argmax(proba: np.ndarray, list_classes: list, **kwargs):
