+++ gabarit/template_nlp/nlp_project/package_name/models_training/model_class.py
@@ -36,7 +36,7 @@ from typing import Optional, List, Dict, Any, Tuple, Union
 from sklearn.metrics import (accuracy_score, confusion_matrix, f1_score, multilabel_confusion_matrix,
                              precision_score, recall_score)
 
-from {{package_name}} import utils
+from .. import utils
 
 sns.set(style="darkgrid")
