+++ gabarit/template_nlp/nlp_project/package_name/models_training/models_tensorflow/model_tfidf_dense.py
@@ -34,9 +34,8 @@ from tensorflow.keras.models import Model, Sequential
 from tensorflow.keras.models import load_model as load_model_keras
 from tensorflow.keras.layers import ELU, BatchNormalization, Dense, Dropout
 
-from {{package_name}} import utils
-from {{package_name}}.models_training import utils_deep_keras
-from {{package_name}}.models_training.model_keras import ModelKeras
+from . import utils_deep_keras
+from .model_keras import ModelKeras
 
 
 sns.set(style="darkgrid")
